Update README.md
Leticia's Portfolio

During the creation of my portfolio, I applied the concepts learned in the course to develop a structured and organized project. To ensure good organization, I created separate folders for the HTML and CSS files.

In the HTML folder, I structured the code in a clear manner, defining the document structure and organizing each section of my portfolio. I created a summary section to introduce myself, highlighting my academic background and career transition into web development. The hobbies and interests section showcases my passions and personal experiences, including activities such as gymnastics, parkour, yoga, and the pursuit of self-discovery. Additionally, I included a contact section with a form for visitors to get in touch with me.

When developing the form, I referred to websites and video tutorials to learn how to correctly implement the necessary elements. I followed best practices to create text fields, an email field, a message field, and a submit button.

In the CSS folder, I focused on styling my portfolio by working with colors and sizes. Following the principles of color psychology, I chose neutral colors that provide a clean and pleasant visual appearance to the site. Throughout the week, I experimented with different color combinations and adjusted the sizes of elements to ensure visual harmony.

This project allowed me to apply the knowledge gained in the course, resulting in a well-structured, aesthetically pleasing, and functional portfolio. I am excited to share my learning journey and projects through this portfolio and invite you to explore and learn more about me.

Welcome to Leticia's Portfolio!